Add files via upload
In this case scenario I look into the banks customer data and perform EDA and data preprocessing, aim is  to build a model to predict the probability of a certain customer churning or remaining in the same financial institution. The corresponding attributes have been indicated and the KPP well selected with respect to how they affect customer churn.